Refactoring visualization components
