+++ src/components/Visualizations/Graphs/BarGraph.tsx
@@ -2,7 +2,7 @@ import { BarDatum, ResponsiveBar } from '@nivo/bar';
 import { ReactElement, useState } from 'react';
 import { Box, Radio, RadioGroup, Stack } from '@chakra-ui/react';
 
-export interface IMetricsGraphProps {
+export interface IBarGraphProps {
   data: BarDatum[];
   indexBy: string;
   keys: string[];
@@ -11,7 +11,7 @@ export interface IMetricsGraphProps {
   showGroupOptions?: boolean;
 }
 
-export const MetricsGraph = (props: IMetricsGraphProps): ReactElement => {
+export const BarGraph = (props: IBarGraphProps): ReactElement => {
   const { data, indexBy, keys, ticks, showLegend = true, showGroupOptions = true } = props;
   const [groupType, setGroupType] = useState('stacked');
