+++ src/components/Visualizations/Graphs/index.ts
@@ -1,1 +1,2 @@
+export * from './BarGraph';
 export * from './LineGraph';
