+++ src/components/Visualizations/Tables/index.ts
@@ -1,8 +1,4 @@
 export * from './CitationsTable';
-export * from './IndicesGraph';
 export * from './IndicesTable';
-export * from './Metrics';
-export * from './MetricsGraph';
 export * from './PapersTable';
 export * from './ReadsTable';
-export * from './types';
