+++ src/components/Visualizations/index.tsx
@@ -1,5 +1,8 @@
 export * from './Containers';
 export * from './GraphPanes';
 export * from './Graphs';
+export * from './Metrics';
+export * from './Tables';
+export * from './types';
 export * from './utils';
 export * from './VisualizationTabs';
