+++ src/components/__stories__/AllSearchTermsDropdown.stories.tsx
@@ -1,4 +1,3 @@
-import { Box } from '@chakra-ui/react';
 import { IAllSearchTermsDropdown, AllSearchTermsDropdown } from '@components';
 import { Meta, Story } from '@storybook/react';
