+++ src/components/__stories__/CitationsTable.stories.tsx
@@ -3,7 +3,7 @@ import { citationsTableData } from '@components/__mocks__/data';
 import { Meta, Story } from '@storybook/react';
 
 const meta: Meta = {
-  title: 'Metrics/CitationsTable',
+  title: 'Visualizations/tables/CitationsTable',
   component: CitationsTable,
 };
 
@@ -11,10 +11,10 @@ export default meta;
 
 const Template: Story<ICitationsTableProps> = (args) => <CitationsTable {...args} />;
 
-export const Default = Template.bind({}) ;
+export const Default = Template.bind({});
 
 Default.args = { data: citationsTableData, isAbstract: false };
 
-export const Abstract = Template.bind({}) ;
+export const Abstract = Template.bind({});
 
 Abstract.args = { data: citationsTableData, isAbstract: true };
