+++ src/components/__stories__/Metrics.stories.tsx
@@ -3,7 +3,7 @@ import { metrics } from '@components/__mocks__/data';
 import { Meta, Story } from '@storybook/react';
 
 const meta: Meta = {
-  title: 'Metrics/Metrics',
+  title: 'Visualizations/Metrics',
   component: Metrics,
 };
 
@@ -11,10 +11,10 @@ export default meta;
 
 const Template: Story<IMetricsProps> = (args) => <Metrics {...args} />;
 
-export const Default = Template.bind({}) ;
+export const Default = Template.bind({});
 
 Default.args = { metrics, isAbstract: false };
 
-export const Abstract = Template.bind({}) ;
+export const Abstract = Template.bind({});
 
 Abstract.args = { metrics, isAbstract: true };
