+++ src/components/__stories__/ReadsTable.stories.tsx
@@ -3,7 +3,7 @@ import { readsTableData } from '@components/__mocks__/data';
 import { Meta, Story } from '@storybook/react';
 
 const meta: Meta = {
-  title: 'Metrics/ReadsTable',
+  title: 'Visualizations/tables/ReadsTable',
   component: ReadsTable,
 };
 
@@ -11,10 +11,10 @@ export default meta;
 
 const Template: Story<IReadsTableProps> = (args) => <ReadsTable {...args} />;
 
-export const Default = Template.bind({}) ;
+export const Default = Template.bind({});
 
 Default.args = { data: readsTableData, isAbstract: false };
 
-export const Abstract = Template.bind({}) ;
+export const Abstract = Template.bind({});
 
 Abstract.args = { data: readsTableData, isAbstract: true };
